refactor(server): split between utils and helprs
